+++ trunk/CXX/Demo/Python3/range.cxx
@@ -123,7 +123,7 @@ Py::Object range::repr()
     return Py::String( asString() );
 }
 
-int range::sequence_length()
+Py_ssize_t range::sequence_length()
 {
     return static_cast<int>( length() );
 }
